+++ scripts/docs.ps1
@@ -1,3 +1,6 @@
+# gen docs, requires python3.x and install dependencies:
+#  pip install -r docs/requirements.tx
+# usage: ./scripts/docs.ps1 '3.39.11:v3.39.11,3.39.12:3.39.x'
 $rel_str = $args[0]
 
 mkdocs --version
@@ -18,7 +21,7 @@ foreach($rel in $rel_arr) {
     # ver:tag
     $info = ($rel -split ':')
     $docs_ver = $info[0]
-    $docs_tag = if ($info.Count -ge 2) { $info[1] } else {$info[0]}
+    $docs_tag = if ($info.Count -ge 2) { $info[1] } else { "v$($info[0])" }
     git checkout "$docs_tag"
     mike deploy $docs_ver
 }
@@ -26,4 +29,4 @@ foreach($rel in $rel_arr) {
 git checkout dev
 mike deploy latest
 mike list
-# mike set-default $docs_ver --push
+mike set-default $docs_ver --push